[NFC] preprocess.cpp: create a separate file with all related functions
